requirements.txt
hydra-core==1.3.2
coloredlogs==15.0.1
numpy==1.26.4
tqdm==4.66.2
wandb>=0.16.4
pandas>=2.2.2
omegaconf==2.3.0
matplotlib
transformers
rotary-embedding-torch
torchtune